register and login user

+++ src/main/java/com/nasya/restapi/entity/User.java
@@ -31,7 +31,7 @@ public class User {
     private String token;
 
     @Column(name = "expired_at")
-    private BigInteger tokenExpiredAt;
+    private Long tokenExpiredAt;
 
     @OneToMany(mappedBy = "user")
     private List<Contact> contacts;
